+++ ujeongri2/src/ujeongri/com/gh/main/run/Run.java
@@ -2,11 +2,12 @@ package ujeongri.com.gh.main.run;
 
 import java.awt.CardLayout;
 
+import ujeongri.com.gh.game.view.PlayerView;
+import ujeongri.com.gh.game.view.ShopView;
 import ujeongri.com.gh.main.view.ButtonMainView;
 import ujeongri.com.gh.main.view.LoginView;
 import ujeongri.com.gh.main.view.MainView;
-import ujeongri.com.gh.main.view.PlayerView;
-import ujeongri.com.gh.main.view.ShopView;
+import ujeongri.com.gh.main.view.TestView;
 
 public class Run {
 
@@ -16,14 +17,8 @@ public class Run {
 //		LoginView lv = new LoginView();
 		ShopView sv = new ShopView();
 //		PlayerView pv = new PlayerView();
+//		TestView tv = new TestView();
 		
 	
-
-		
-		
-
 	}
-	
-	
-
 }
